+++ tests/training/test_geometric_median.py
@@ -1,0 +1,1 @@
+# Placeholder for geometric median tests
