--- tests/training/test_sae_training.py
@@ -1,382 +1,0 @@
-from typing import Any
-
-import einops
-import pytest
-import torch
-from datasets import Dataset
-from transformer_lens import HookedTransformer
-
-from sae_lens.config import LanguageModelSAERunnerConfig
-from sae_lens.saes.jumprelu_sae import JumpReLU
-from sae_lens.saes.sae import TrainingSAE, TrainStepInput
-from sae_lens.training.activations_store import ActivationsStore
-from sae_lens.training.sae_trainer import SAETrainer
-from tests.helpers import build_sae_cfg
-
-
-# Define a new fixture for different configurations
-@pytest.fixture(
-    params=[
-        {
-            "model_name": "tiny-stories-1M",
-            "dataset_path": "roneneldan/TinyStories",
-            "hook_name": "blocks.1.hook_resid_pre",
-            "hook_layer": 1,
-            "d_in": 64,
-        },
-        {
-            "model_name": "tiny-stories-1M",
-            "dataset_path": "roneneldan/TinyStories",
-            "hook_name": "blocks.1.hook_resid_pre",
-            "hook_layer": 1,
-            "d_in": 64,
-            "normalize_sae_decoder": False,
-            "scale_sparsity_penalty_by_decoder_norm": True,
-        },
-        {
-            "model_name": "tiny-stories-1M",
-            "dataset_path": "apollo-research/roneneldan-TinyStories-tokenizer-gpt2",
-            "hook_name": "blocks.1.hook_resid_pre",
-            "hook_layer": 1,
-            "d_in": 64,
-        },
-        {
-            "model_name": "tiny-stories-1M",
-            "dataset_path": "apollo-research/roneneldan-TinyStories-tokenizer-gpt2",
-            "hook_name": "blocks.1.hook_resid_pre",
-            "hook_layer": 1,
-            "d_in": 64,
-            "normalize_activations": "constant_norm_rescale",
-        },
-        {
-            "model_name": "tiny-stories-1M",
-            "dataset_path": "roneneldan/TinyStories",
-            "hook_name": "blocks.1.attn.hook_z",
-            "hook_layer": 1,
-            "d_in": 64,
-        },
-    ],
-    ids=[
-        "tiny-stories-1M-resid-pre",
-        "tiny-stories-1M-resid-pre-L1-W-dec-Norm",
-        "tiny-stories-1M-resid-pre-pretokenized",
-        "tiny-stories-1M-resid-pre-pretokenized-norm-rescale",
-        "tiny-stories-1M-hook-z",
-    ],
-)
-def cfg(request: pytest.FixtureRequest):
-    """
-    Pytest fixture to create a mock instance of LanguageModelSAERunnerConfig.
-    """
-    params = request.param
-    return build_sae_cfg(**params)
-
-
-@pytest.fixture
-def training_sae(cfg: Any):
-    """
-    Pytest fixture to create a mock instance of SparseAutoencoder.
-    """
-    return TrainingSAE.from_dict(cfg.get_training_sae_cfg_dict())
-
-
-@pytest.fixture
-def activation_store(model: HookedTransformer, cfg: LanguageModelSAERunnerConfig):
-    return ActivationsStore.from_config(
-        model, cfg, override_dataset=Dataset.from_list([{"text": "hello world"}] * 2000)
-    )
-
-
-@pytest.fixture
-def model(cfg: LanguageModelSAERunnerConfig):
-    return HookedTransformer.from_pretrained(cfg.model_name, device="cpu")
-
-
-# todo: remove the need for this fixture
-@pytest.fixture
-def trainer(
-    cfg: LanguageModelSAERunnerConfig,
-    training_sae: TrainingSAE,
-    model: HookedTransformer,
-    activation_store: ActivationsStore,
-):
-    return SAETrainer(
-        model=model,
-        sae=training_sae,
-        activation_store=activation_store,
-        save_checkpoint_fn=lambda *args, **kwargs: None,  # noqa: ARG005
-        cfg=cfg,
-    )
-
-
-# TODO: DECIDE IF WE ARE KEEPING ENCODE AND DECODE METHODS
-
-# def test_sparse_autoencoder_encode(training_sae: TrainingSparseAutoencoder):
-#     batch_size = 32
-#     d_in = training_sae.d_in
-#     d_sae = training_sae.d_sae
-
-#     x = torch.randn(batch_size, d_in)
-#     feature_acts1 = training_sae.encode(x)
-#     _, cache = training_sae.run_with_cache(x, names_filter="hook_sae_acts_post")
-#     feature_acts2 = cache["hook_sae_acts_post"]
-
-#     # Check shape
-#     assert feature_acts2.shape == (batch_size, d_sae)
-
-#     # Check values
-#     assert torch.allclose(feature_acts1, feature_acts2)
-
-
-# def test_sparse_autoencoder_decode(training_sae: TrainingSparseAutoencoder):
-#     batch_size = 32
-#     d_in = training_sae.d_in
-
-#     x = torch.randn(batch_size, d_in)
-#     sae_out1 = training_sae(x)
-
-#     assert sae_out1.shape == x.shape
-#     assert torch.allclose(sae_out1, sae_out2)
-
-
-def test_sae_forward(training_sae: TrainingSAE):
-    batch_size = 32
-    d_in = training_sae.cfg.d_in
-    d_sae = training_sae.cfg.d_sae
-
-    x = torch.randn(batch_size, d_in)
-    # print(f"\nTest input shape: {x.shape}")
-    # print(f"SAE d_in: {training_sae.cfg.d_in}, d_sae: {training_sae.cfg.d_sae}")
-    # print(f"SAE hook_name: {training_sae.cfg.hook_name}")
-    # print(f"SAE hook_z_reshaping_mode: {getattr(training_sae, 'hook_z_reshaping_mode', False)}")
-
-    train_step_output = training_sae.training_forward_pass(
-        step_input=TrainStepInput(
-            sae_in=x,
-            current_l1_coefficient=training_sae.cfg.l1_coefficient,
-            dead_neuron_mask=None,
-        )
-    )
-
-    assert train_step_output.sae_out.shape == (batch_size, d_in)
-    assert train_step_output.feature_acts.shape == (batch_size, d_sae)
-    assert (
-        pytest.approx(train_step_output.loss.detach(), rel=1e-3)
-        == (train_step_output.losses["mse_loss"] + train_step_output.losses["l1_loss"])
-        .detach()  # type: ignore
-        .cpu()
-        .numpy()
-    )
-
-    expected_mse_loss = (
-        (torch.pow((train_step_output.sae_out - x.float()), 2))
-        .sum(dim=-1)
-        .mean()
-        .detach()
-        .float()
-    )
-
-    assert (
-        pytest.approx(train_step_output.losses["mse_loss"].item()) == expected_mse_loss  # type: ignore
-    )
-
-    if not training_sae.cfg.scale_sparsity_penalty_by_decoder_norm:
-        expected_l1_loss = train_step_output.feature_acts.sum(dim=1).mean(dim=(0,))
-    else:
-        expected_l1_loss = (
-            (train_step_output.feature_acts * training_sae.W_dec.norm(dim=1))
-            .norm(dim=1, p=1)
-            .mean()
-        )
-    assert (
-        pytest.approx(train_step_output.losses["l1_loss"].item(), rel=1e-3)  # type: ignore
-        == training_sae.cfg.l1_coefficient * expected_l1_loss.detach().float()
-    )
-
-
-def test_sae_forward_with_mse_loss_norm(
-    training_sae: TrainingSAE,
-):
-    # change the confgi and ensure the mse loss is calculated correctly
-    training_sae.cfg.mse_loss_normalization = "dense_batch"
-    training_sae.mse_loss_fn = training_sae._get_mse_loss_fn()
-
-    batch_size = 32
-    d_in = training_sae.cfg.d_in
-    d_sae = training_sae.cfg.d_sae
-
-    x = torch.randn(batch_size, d_in)
-    train_step_output = training_sae.training_forward_pass(
-        step_input=TrainStepInput(
-            sae_in=x,
-            current_l1_coefficient=training_sae.cfg.l1_coefficient,
-            dead_neuron_mask=None,
-        )
-    )
-
-    assert train_step_output.sae_out.shape == (batch_size, d_in)
-    assert train_step_output.feature_acts.shape == (batch_size, d_sae)
-    assert "ghost_grad_loss" not in train_step_output.losses
-
-    x_centred = x - x.mean(dim=0, keepdim=True)
-    expected_mse_loss = (
-        (
-            torch.nn.functional.mse_loss(train_step_output.sae_out, x, reduction="none")
-            / (1e-6 + x_centred.norm(dim=-1, keepdim=True))
-        )
-        .sum(dim=-1)
-        .mean()
-        .detach()
-        .item()
-    )
-
-    assert (
-        pytest.approx(train_step_output.losses["mse_loss"].item()) == expected_mse_loss  # type: ignore
-    )
-
-    assert (
-        pytest.approx(train_step_output.loss.detach(), rel=1e-3)
-        == (train_step_output.losses["mse_loss"] + train_step_output.losses["l1_loss"])
-        .detach()  # type: ignore
-        .numpy()
-    )
-
-    if not training_sae.cfg.scale_sparsity_penalty_by_decoder_norm:
-        expected_l1_loss = train_step_output.feature_acts.sum(dim=1).mean(dim=(0,))
-    else:
-        expected_l1_loss = (
-            (train_step_output.feature_acts * training_sae.W_dec.norm(dim=1))
-            .norm(dim=1, p=1)
-            .mean()
-        )
-    assert (
-        pytest.approx(train_step_output.losses["l1_loss"].item(), rel=1e-3)  # type: ignore
-        == training_sae.cfg.l1_coefficient * expected_l1_loss.detach().float()
-    )
-
-
-def test_per_item_mse_loss_with_norm_matches_original_implementation(
-    training_sae: TrainingSAE,
-) -> None:
-    training_sae.cfg.mse_loss_normalization = "dense_batch"
-    training_sae.mse_loss_fn = training_sae._get_mse_loss_fn()
-
-    input = torch.randn(3, 2)
-    target = torch.randn(3, 2)
-    target_centered = target - target.mean(dim=0, keepdim=True)
-    orig_impl_res = (
-        torch.pow((input - target.float()), 2)
-        / (target_centered**2).sum(dim=-1, keepdim=True).sqrt()
-    )
-    sae_res = training_sae.mse_loss_fn(
-        input,
-        target,
-    )
-    assert torch.allclose(orig_impl_res, sae_res, atol=1e-5)
-
-
-def test_SparseAutoencoder_forward_can_add_noise_to_hidden_pre() -> None:
-    clean_cfg = build_sae_cfg(d_in=2, d_sae=4, noise_scale=0)
-    noisy_cfg = build_sae_cfg(d_in=2, d_sae=4, noise_scale=100)
-    clean_sae = TrainingSAE.from_dict(clean_cfg.get_training_sae_cfg_dict())
-    noisy_sae = TrainingSAE.from_dict(noisy_cfg.get_training_sae_cfg_dict())
-
-    input = torch.randn(3, 2)
-
-    clean_output1 = clean_sae.forward(input)
-    clean_output2 = clean_sae.forward(input)
-    noisy_output1 = noisy_sae.forward(input)
-    noisy_output2 = noisy_sae.forward(input)
-
-    # with no noise, the outputs should be identical
-    assert torch.allclose(clean_output1, clean_output2)
-    # noisy outputs should be different
-    assert not torch.allclose(noisy_output1, noisy_output2)
-    assert not torch.allclose(clean_output1, noisy_output1)
-
-
-def test_SparseAutoencoder_remove_gradient_parallel_to_decoder_directions(
-    training_sae: TrainingSAE,
-) -> None:
-    if not training_sae.cfg.normalize_sae_decoder:
-        pytest.skip("Test only applies when decoder is not normalized")
-    sae = training_sae
-    orig_grad = torch.randn_like(sae.W_dec)
-    orig_W_dec = sae.W_dec.clone()
-    sae.W_dec.grad = orig_grad.clone()
-    sae.remove_gradient_parallel_to_decoder_directions()
-
-    # check that the gradient is orthogonal to the decoder directions
-    parallel_component = einops.einsum(
-        sae.W_dec.grad,
-        sae.W_dec.data,
-        "d_sae d_in, d_sae d_in -> d_sae",
-    )
-
-    assert torch.allclose(
-        parallel_component, torch.zeros_like(parallel_component), atol=1e-5
-    )
-    # the decoder weights should not have changed
-    assert torch.allclose(sae.W_dec, orig_W_dec)
-
-    # the gradient delta should align with the decoder directions
-    grad_delta = orig_grad - sae.W_dec.grad
-    assert torch.nn.functional.cosine_similarity(
-        sae.W_dec.detach(), grad_delta, dim=1
-    ).abs() == pytest.approx(1.0, abs=1e-3)
-
-
-def test_SparseAutoencoder_set_decoder_norm_to_unit_norm(
-    trainer: SAETrainer,
-) -> None:
-    if not trainer.cfg.normalize_sae_decoder:
-        pytest.skip("Test only applies when decoder is not normalized")
-
-    sae = trainer.sae
-    sae.W_dec.data = 20 * torch.randn_like(sae.W_dec)
-    sae.set_decoder_norm_to_unit_norm()
-    assert torch.allclose(
-        torch.norm(sae.W_dec, dim=1), torch.ones_like(sae.W_dec[:, 0])
-    )
-
-
-def test_jumprelu_forward():
-    x = torch.tensor([-1.0, 0.5, 1.5, 2.5])
-    threshold = torch.tensor(1.0)
-    expected_output = torch.tensor([0.0, 0.0, 1.5, 2.5])
-    output = JumpReLU.apply(x, threshold, 0.001)
-    assert torch.allclose(output, expected_output)  # type: ignore
-
-
-def test_jumprelu_backward():
-    x = torch.tensor([-1.0, 0.5, 1.5, 2.5], requires_grad=True)
-    threshold = torch.tensor(1.0)
-    output = JumpReLU.apply(x, threshold, 0.001)
-    output.sum().backward()  # type: ignore
-    expected_grad_x = torch.tensor([0.0, 0.0, 1.0, 1.0])
-    assert torch.allclose(x.grad, expected_grad_x)  # type: ignore
-
-
-def test_jumprelu_backward_with_threshold_grad():
-    bandwidth = 1e-3
-    threshold_value = 1.0
-    threshold = torch.tensor(threshold_value, requires_grad=True)
-
-    epsilon = 1e-8
-    x = torch.tensor(
-        [
-            threshold_value - bandwidth / 2 + epsilon,
-            threshold_value + bandwidth / 2 - epsilon,
-        ],
-        requires_grad=True,
-    )
-
-    output = JumpReLU.apply(x, threshold, bandwidth)
-    output.sum().backward()  # type: ignore
-
-    expected_grad_x = torch.tensor([0.0, 1.0])
-    expected_grad_threshold = torch.tensor(-2000.0)
-
-    atol = 1e-2
-    assert torch.allclose(x.grad, expected_grad_x, atol=atol)  # type: ignore
-    assert torch.isclose(threshold.grad, expected_grad_threshold, atol=atol)  # type: ignore
